phase-2/cloud-function/requirements.txt
google-genai
google-cloud-bigquery
